(load_osxfuse) Add support for KextManager API and replace Gestalt
KextManager API has been instroduced in Mac OS X 10.6 and includes the
following functions for loading and unloading kernel extensions:

* KextManagerLoadKextWithURL
* KextManagerUnloadKextWithIdentifier (available in 10.7 and later)

Use those functions instead of the system tools kextload or kextunload in case
they are avaliable, i.e. we are running on Mac OS X 10.6 or later.

Gestalt has been deprecated in OS X 10.8, therefore we are using
/System/Library/CoreServices/SystemVersion.plist to determine which version of
Mac OS X we are running on.